Extract cluster state task classes from ShardStateAction
