+++ server/src/main/java/org/elasticsearch/indices/recovery/package-info.java
@@ -169,13 +169,13 @@
  *         recovery. This leads to a call to {@link org.elasticsearch.index.shard.IndexShard#postRecovery} which moves the recovery state
  *         to stage {@link org.elasticsearch.indices.recovery.RecoveryState.Stage#DONE}, triggers a refresh of the shard and moves the
  *         shard to state {@link org.elasticsearch.index.shard.IndexShardState#POST_RECOVERY}. Finally, the recovery target will then
- *         send a {@link org.elasticsearch.cluster.action.shard.ShardStateAction.StartedShardEntry} transport message to master to inform
+ *         send a {@link org.elasticsearch.cluster.action.shard.StartedShardEntry} transport message to master to inform
  *         it about the successful start of the shard.
  *     </li>
  *     <li>
  *         After receiving the {@code StartedShardEntry}, master will then update the cluster state to reflect the state of the now fully
  *         recovered recovery target by executing the
- *         {@link org.elasticsearch.cluster.action.shard.ShardStateAction.ShardStartedClusterStateTaskExecutor}. The resulting cluster
+ *         {@link org.elasticsearch.cluster.action.shard.ShardStartedTaskExecutor}. The resulting cluster
  *         state update is then observed by {@link org.elasticsearch.index.shard.IndexShard#updateShardState} which updates the shard state
  *         on the target node to {@link org.elasticsearch.index.shard.IndexShardState#STARTED} thus completing the peer recovery.
  *     </li>
