+++ server/src/test/java/org/elasticsearch/indices/cluster/ClusterStateChanges.java
@@ -42,10 +42,10 @@ import org.elasticsearch.cluster.ClusterStateTaskExecutor;
 import org.elasticsearch.cluster.ClusterStateTaskListener;
 import org.elasticsearch.cluster.EmptyClusterInfoService;
 import org.elasticsearch.cluster.TestShardRoutingRoleStrategies;
-import org.elasticsearch.cluster.action.shard.ShardStateAction;
-import org.elasticsearch.cluster.action.shard.ShardStateAction.FailedShardUpdateTask;
-import org.elasticsearch.cluster.action.shard.ShardStateAction.StartedShardEntry;
-import org.elasticsearch.cluster.action.shard.ShardStateAction.StartedShardUpdateTask;
+import org.elasticsearch.cluster.action.shard.FailedShardEntry;
+import org.elasticsearch.cluster.action.shard.ShardFailedTaskExecutor;
+import org.elasticsearch.cluster.action.shard.ShardStartedTaskExecutor;
+import org.elasticsearch.cluster.action.shard.StartedShardEntry;
 import org.elasticsearch.cluster.block.ClusterBlock;
 import org.elasticsearch.cluster.coordination.JoinReason;
 import org.elasticsearch.cluster.coordination.JoinTask;
@@ -140,8 +140,8 @@ public class ClusterStateChanges {
     private final AllocationService allocationService;
     private final ClusterService clusterService;
     private final FeatureService featureService;
-    private final ShardStateAction.ShardFailedClusterStateTaskExecutor shardFailedClusterStateTaskExecutor;
-    private final ShardStateAction.ShardStartedClusterStateTaskExecutor shardStartedClusterStateTaskExecutor;
+    private final ShardFailedTaskExecutor shardFailedClusterStateTaskExecutor;
+    private final ShardStartedTaskExecutor shardStartedTaskExecutor;
 
     // transport actions
     private final TransportOpenIndexAction transportOpenIndexAction;
@@ -171,12 +171,8 @@ public class ClusterStateChanges {
             EmptySnapshotsInfoService.INSTANCE,
             TestShardRoutingRoleStrategies.DEFAULT_ROLE_ONLY
         );
-        shardFailedClusterStateTaskExecutor = new ShardStateAction.ShardFailedClusterStateTaskExecutor(allocationService, null);
-        shardStartedClusterStateTaskExecutor = new ShardStateAction.ShardStartedClusterStateTaskExecutor(
-            clusterSettings,
-            allocationService,
-            null
-        );
+        shardFailedClusterStateTaskExecutor = new ShardFailedTaskExecutor(allocationService, null);
+        shardStartedTaskExecutor = new ShardStartedTaskExecutor(clusterSettings, allocationService, null);
         ActionFilters actionFilters = new ActionFilters(Collections.emptySet());
         IndexNameExpressionResolver indexNameExpressionResolver = TestIndexNameExpressionResolver.newInstance();
         DestructiveOperations destructiveOperations = new DestructiveOperations(SETTINGS, clusterSettings);
@@ -466,10 +462,10 @@ public class ClusterStateChanges {
     }
 
     public ClusterState applyFailedShards(ClusterState clusterState, List<FailedShard> failedShards) {
-        List<FailedShardUpdateTask> entries = failedShards.stream()
+        List<ShardFailedTaskExecutor.Task> entries = failedShards.stream()
             .map(
-                failedShard -> new FailedShardUpdateTask(
-                    new ShardStateAction.FailedShardEntry(
+                failedShard -> new ShardFailedTaskExecutor.Task(
+                    new FailedShardEntry(
                         failedShard.routingEntry().shardId(),
                         failedShard.routingEntry().allocationId().getId(),
                         0L,
@@ -494,12 +490,12 @@ public class ClusterStateChanges {
 
     public ClusterState applyStartedShards(ClusterState clusterState, Map<ShardRouting, Long> startedShards) {
         return runTasks(
-            shardStartedClusterStateTaskExecutor,
+            shardStartedTaskExecutor,
             clusterState,
             startedShards.entrySet()
                 .stream()
                 .map(
-                    e -> new StartedShardUpdateTask(
+                    e -> new ShardStartedTaskExecutor.Task(
                         new StartedShardEntry(
                             e.getKey().shardId(),
                             e.getKey().allocationId().getId(),
