+++ x-pack/plugin/stateless/src/internalClusterTest/java/org/elasticsearch/xpack/stateless/recovery/StatelessDirectRecoveryCancellationIT.java
@@ -14,6 +14,7 @@ import org.elasticsearch.action.admin.cluster.reroute.ClusterRerouteUtils;
 import org.elasticsearch.action.admin.cluster.reroute.TransportClusterRerouteAction;
 import org.elasticsearch.action.admin.indices.ResizeIndexTestUtils;
 import org.elasticsearch.action.admin.indices.shrink.ResizeType;
+import org.elasticsearch.cluster.action.shard.FailedShardEntry;
 import org.elasticsearch.cluster.action.shard.ShardStateAction;
 import org.elasticsearch.cluster.routing.RecoverySource;
 import org.elasticsearch.cluster.routing.ShardRouting;
@@ -462,7 +463,7 @@ public class StatelessDirectRecoveryCancellationIT extends AbstractStatelessPlug
         final var shardFailureReceivedLatch = new CountDownLatch(1);
         MockTransportService.getInstance(node)
             .addRequestHandlingBehavior(ShardStateAction.SHARD_FAILED_ACTION_NAME, (handler, request, channel, task) -> {
-                if (request instanceof ShardStateAction.FailedShardEntry failedShard) {
+                if (request instanceof FailedShardEntry failedShard) {
                     if (failedShard.getShardId().equals(shardId)
                         && ExceptionsHelper.unwrap(failedShard.getFailure(), RecoveryCancelledException.class) != null) {
                         shardFailureReceivedLatch.countDown();
